allow to match results of attestation in tests
